+++ CHANGELOG.md
@@ -2,10 +2,12 @@
 
 ## [0.9.1] - Unreleased
 ### New
+- Add definition for the `Preselection` element.
 ### Changed
 - Downloading: only download subtitles when `fetch_subtitles()` has been called on DashDownloader
   (from @sleepycatcoding).
 - Add derived PartialEq to data structures to allow for comparison.
+
 
 ## [0.9.0] - 2023-05-10
 ### New